逻辑对了，并且能对点，但是追踪效果不好

+++ Lib/Src/StepMotor.c
@@ -1,6 +1,7 @@
 #include "StepMotor.h"
 #include "stdlib.h"
 #include "EMM_V5.h"
+#include <stdint.h>
 
 void Motor_Set_Zero(StepMotor_Driver *driver)
 {
@@ -10,9 +11,9 @@ void Motor_Set_Zero(StepMotor_Driver *driver)
 void Motor_Move(StepMotor_Driver *driver, float val)
 {
     if(val>=0)
-    Emm_V5_Vel_Control(driver,1,val,0,0);
+    Emm_V5_Vel_Control(driver,1,(int16_t)val,0,0);
     else 
-    Emm_V5_Vel_Control(driver,0,(-val),0,0);
+    Emm_V5_Vel_Control(driver,0,(int16_t)(-val),0,0);
 }
 
 void Motor_Stop(StepMotor_Driver *driver)
